+++ docs/component---src-pages-index-tsx-096fe8fc29657969ff5b.js
@@ -1,2 +1,0 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[17],{QeBL:function(e,t,n){"use strict";n.r(t);var a=n("q1tI"),r=n.n(a),c=n("9Dj+"),o=n("H8eV");t.default=function(){return r.a.createElement(c.a,null,r.a.createElement(o.a,{title:"Home",keywords:["infoSum","charts","react"],description:""}))}}}]);
-//# sourceMappingURL=component---src-pages-index-tsx-096fe8fc29657969ff5b.js.map